KAN-162 refactor: 상품패키지구조 변경-상품 DTO, SERVICE 패키지 고객, 회원단위로 재설정
