+++ src/main/java/com/yeonieum/productservice/domain/product/dto/memberservice/RetrieveAdvertisementProductResponseDto.java
@@ -1,9 +1,8 @@
-package com.yeonieum.productservice.domain.product.dto;
+package com.yeonieum.productservice.domain.product.dto.memberservice;
 
 import com.yeonieum.productservice.global.enums.ActiveStatus;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 public class RetrieveAdvertisementProductResponseDto {
     Long productId;
